fix(Credits): stop rounding half-credit courses down to zero
NTUST issues 0.5-credit courses. Credits were an Int end to end, and both
parsers reached them through toIntOrNull() / ^(\d+)$ — so every one of
those courses arrived as 0: missing from the class table's total, blank on
the transcript row, and weightless in the GPA. The backend already sends a
float and SyncApiClient truncated it on the way back in.

- Course.credits and CourseGrade.credits are Float; the transcript's credit
  patterns take an optional fraction in every bracket form, so a half credit
  keeps its credit type instead of falling through to UNKNOWN
- CreditBreakdown follows, since the school's own summary table adds up the
  same halves
- Float.formatCredits() prints "3" and "0.5" rather than "3.0", locale-aware
- no DataMigration step: Float is a JVM primitive, so Gson's Unsafe path
  leaves 0.0f rather than null, and JSON has one number type — a cache
  written as "credits":3 still reads back as 3.0f. HalfCreditTest pins that

+++ app/src/main/java/org/ntust/app/tigerduck/data/model/ScoreReport.kt
@@ -35,7 +35,8 @@ data class CourseGrade(
     val term: String,
     val code: String,
     val name: String,
-    val credits: Int? = null,
+    /** Fractional: NTUST issues half credits. */
+    val credits: Float? = null,
     val creditType: CreditType = CreditType.UNKNOWN,
     val grade: String = "",
     val status: GradeStatus = GradeStatus.UNKNOWN,
@@ -88,11 +89,11 @@ data class CreditSummary(
 }
 
 data class CreditBreakdown(
-    val inPerson: Int,
-    val distance: Int,
-    val total: Int
+    val inPerson: Float,
+    val distance: Float,
+    val total: Float
 ) {
     companion object {
-        val ZERO = CreditBreakdown(0, 0, 0)
+        val ZERO = CreditBreakdown(0f, 0f, 0f)
     }
 }
